Add tests for Task 2
