+++ prac5/template_makefile
@@ -1,16 +1,17 @@
 # To be configured by configure.sh:
 # 	black_fitch_path=/path/to/black-fitch/prac5
 
-task1_test_files_local=JavaFitchRunner.java UnitTest.java tests_Task1.java
-
+# Task 1 and 2
+task1_test_files_local=JavaFitchRunner.java UnitTest.java tests_All.java
 task1_source_files=$(task1_test_files_local) TreapOperations.java Node.java
 
 task1_test_files_black_fitch=$(black_fitch_path)/JavaFitchRunner.java $(black_fitch_path)/UnitTest.java \
-$(black_fitch_path)/tests_Task1.java
+$(black_fitch_path)/tests_All.java
 
 task1_main=JavaFitchRunner
 
-task1:
+
+all:
 	@echo "\nCopying Java Fitch..."
 	cp -f $(task1_test_files_black_fitch) .
 	@echo ""
